requirements.txt
# AI Reels Bot - Dependencies
# Python 3.12+

moviepy>=1.0.3
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
playwright>=1.40.0
